test: update fixture to show reactivity issue
