+++ test/fixtures/bridge/layouts/default.vue
@@ -1,0 +1,17 @@
+<template>
+  <div>
+    <Nuxt />
+    {{ route.path }}
+  </div>
+</template>
+
+<script>
+import { useRoute } from '#app'
+export default {
+  setup () {
+    const route = useRoute()
+    console.log(route.path)
+    return { route }
+  }
+}
+</script>
